requirement.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.7.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
